Several actions regarding the stitching of images:
Added dialog for opening a set of images.
Images selected are copied into Fragments folder an added to imagenames
list.
When opening a project, Fragments folder is read and loaded into
imagenames list.
Imagenames list is shown. When deleting an image from the list, it is
deleted from the Fragments Folder.
Added Thread to execute OpenCV stitching executable, output image is
called Full_Image.png on the folder Full_Image.
Full Image is loaded into the preview and the mainApp reference when it
is generated.