Added relative vertical line.
Added mouse event that, when clicked, stores the absolute position of a
point and draws a vertical line on that point. When resizing the image,
this line gets resized and stays in the same image-position.
Still need to do some fixes and use this to mark the Start-End of the
tooth and mark the divisions.