Added Free-Drawing and Start-End point-Deciles.
Added Selection of Start-End points for obtaining the deciles, Drawing
of the Start, end and all the deciles lines and the redimension of those
lines when zooming.

Added Free Draw option, redimension of the free-draw line when zooming.

Working on: 
Improving the quality of the code (bad on PerikymataCountController
right now).
Getting the intensity profile of the free draw line.